+++ lib/presentation/home/view/widgets/compass_pointer/view/compass_pointer.dart
@@ -112,14 +112,17 @@ class _LoadedCompassState extends State<_LoadedCompass> {
 
   @override
   Widget build(BuildContext context) {
-    return Pointer(
-      rotateAngle: widget.state.angle,
-      arcRadius: widget.state.accuracy,
-      radius: 40,
-      foregroundColor: Colors.white,
-      backGroundColor: (widget.state.needCalibration && Platform.isAndroid)
-          ? _color
-          : Colors.grey,
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Pointer(
+        rotateAngle: widget.state.angle,
+        arcRadius: widget.state.accuracy,
+        radius: 40,
+        foregroundColor: Colors.white,
+        backGroundColor: (widget.state.needCalibration && Platform.isAndroid)
+            ? _color
+            : Colors.grey,
+      ),
     );
   }
